refactor: simplify instance context helpers in prompt tests
